More small style fixes, everything looks nice and shiny
