+++ app/components/App.tsx
@@ -8,13 +8,13 @@ require('expedition-app/app/style.scss')
 import {EditorState, QuestType} from '../reducers/StateTypes'
 
 export interface AppStateProps {
+  // TODO Cleanup
   editor: EditorState;
   quest: QuestType;
   scope: any;
 }
 
 export interface AppDispatchProps {
-  playFromCursor: (baseScope: any, editor: EditorState, quest: QuestType) => void;
 }
 
 interface AppProps extends AppStateProps, AppDispatchProps {}
@@ -22,18 +22,6 @@ interface AppProps extends AppStateProps, AppDispatchProps {}
 const App = (props: AppProps): JSX.Element => {
   return (
     <div className="app_root">
-      <div className="app_controls">
-        <FlatButton
-          onTouchTap={(event: any) => props.playFromCursor({}, props.editor, props.quest)}
-          label="Play from Cursor">
-        </FlatButton>
-        {props.editor.bottomPanelShown && 
-        <FlatButton
-          onTouchTap={(event: any) => props.playFromCursor(props.scope, props.editor, props.quest)} 
-          label="Play from Cursor (preserve context)">
-        </FlatButton>
-        }
-      </div>
       <div className="app editor_override">
         <Main/>
       </div>
